clean up code
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@11421 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genometryImpl/src/com/affymetrix/genometryImpl/operator/MeanOperator.java
@@ -25,7 +25,7 @@ public class MeanOperator extends AbstractGraphOperator implements Operator, Ope
 		for (Float f : operands) {
 			total += f.floatValue();
 		}
-		return total / (float)operands.size();
+		return total / operands.size();
 	}
 	
 	@Override
